moved elections into its own module folder
