+++ src/app/app-routing.module.ts
@@ -2,12 +2,6 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { ErrorPageComponent } from '../shared-ng/components/error-page/error-page.component';
 
-import {
-  HomeComponent,
-  VoteComponent,
-  AdminComponent
-} from './routes/routes';
-
 
 const routes: Routes = [
   {
@@ -19,12 +13,8 @@ const routes: Routes = [
     loadChildren: './modules/mask/mask.module#MaskModule'
   },
   {
-    'path': 'vote',
-    component: VoteComponent
-  },
-  {
-    'path': 'admin',
-    component: AdminComponent
+    'path': 'elections',
+    loadChildren: './modules/elections/elections.module#ElectionsModule'
   },
   {
     path: '**',
