+++ src/app/modules/elections/components/admin-ballot-modal/admin-ballot-modal.component.html
