+++ src/app/modules/elections/components/admin-candidates/admin-elections-candidate-row.component.html
