+++ src/app/modules/elections/components/admin-candidates/admin-elections-candidate.component.css
