+++ src/app/modules/elections/components/admin-elections/admin-elections-row.component.html
