+++ src/app/modules/elections/components/admin-elections/admin-elections.component.css
