+++ src/app/modules/elections/components/admin/admin.component.html
