+++ src/app/modules/elections/components/elections.component.ts
@@ -1,0 +1,15 @@
+export * from './home/home.component';
+export * from './vote/vote.component';
+export * from './admin/admin.component';
+export * from './admin-elections/admin-elections.component';
+export * from './admin-positions/admin-positions.component';
+export * from './admin-candidates/admin-elections-candidate-modal.component';
+export * from './admin-ballots/admin-ballots.component';
+export * from './admin-ballot-modal/admin-ballot-modal.component';
+export * from './countdown/countdown.component';
+export * from './vote-form/vote-form.component';
+export * from './multi-position-handler/multi-position-handler.component';
+export * from './district-selection/district-selection.component';
+export * from './navigator/navigator.component';
+
+
