+++ src/app/modules/elections/components/home/home.component.css
