+++ src/app/modules/elections/components/multi-position-handler/multi-position-handler.component.spec.ts
