+++ src/app/modules/elections/components/multi-position-handler/multi-position-handler.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
-import { PageTransitions } from 'src/app/routes/vote/vote.component';
+import { PageTransitions } from 'src/app/modules/elections/components/vote/vote.component';
 import { Election, Position, Vote } from 'src/shared-ng/interfaces/elections';
 
 @Component({
