+++ src/app/modules/elections/components/navigator/navigator.component.html
