+++ src/app/modules/elections/components/navigator/navigator.component.spec.ts
