+++ src/app/modules/elections/components/navigator/navigator.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input, EventEmitter, Output } from '@angular/core';
 
-import { PageTransitions } from 'src/app/routes/vote/vote.component'
+import { PageTransitions } from 'src/app/modules/elections/components/vote/vote.component'
 
 
 @Component({
