+++ src/app/modules/elections/components/vote-form/vote-form.component.spec.ts
