+++ src/app/modules/elections/components/vote/vote.component.spec.ts
