--- src/app/routes/routes.ts
@@ -1,3 +1,0 @@
-export * from './home/home.component';
-export * from './vote/vote.component';
-export * from './admin/admin.component';
