--- src/app/shared/admin/admin.ts
@@ -1,5 +1,0 @@
-export * from './admin-elections/admin-elections.component';
-export * from './admin-positions/admin-positions.component';
-export * from './admin-candidates/admin-elections-candidate-modal.component';
-export * from './admin-ballots/admin-ballots.component';
-export * from './admin-ballot-modal/admin-ballot-modal.component';
