--- src/app/shared/shared.ts
@@ -1,6 +1,0 @@
-export * from './countdown/countdown.component';
-export * from './admin/admin';
-export * from './vote-form/vote-form.component';
-export * from './multi-position-handler/multi-position-handler.component';
-export * from './district-selection/district-selection.component';
-export * from './navigator/navigator.component';
